clean-up of unused code
